Generated code optimizations (#156)
* Opportunistically update a few dependencies

This version of autovalue finally fixes the weird invalid jar issue that would cause some extensions to not run (an issue I had a lot with this project locally). Also uses latest gradle and some misc deps

* Make person have a nullable address for demo purposes

* If property is nullable, check first before writing

This is a minor optimization that saves space on the wire. By not writing null values, we don't have to write their keys either.

* Add test to make sure null nullable props are omitted

* Reuse adapters for types rather than look up duplicate adapters

* Update tests for new optimizations

* Remove nullable property checks

* Update tests after removing nullability checks

* Shorten typenames to simple names

* Update tests

* Fix test