Simplify generics handling by switching to Type[] array for generics handling (#193)
Simplifies the way generics are handled. Before this we were always going through and trying to pass around or unpack a typetoken and mess with the generics, only to raw them out to `Type` under the hood anyway. This basically matches how we do it in `auto-value-moshi` and switches to just accept that `Type[]` directly and also handle that in the generated factory.

Resolves #79 
Resolves #115