Clean up Ignore logic a bit
Relates to #48, but I can't repro any issue described there and think it may just be old. This improves it a bit by not always inverting the serialization checks.

I do plan to just use a shared transient annotation library for this in a separate PR, but this should make switching to that easy