Switch to AutoTransient (#207)
This switches to a shared `@AutoTransient` annotation/artifact for marking fields as effectively transient. Note that this is a slight behavior change too - there is no support for serialization or deserialization-only. This treats it like a regular `transient` modifier, and if a consumer only wants serialization for just one type, they should make their own delegating adapter that only forwards the type they want to support.

Still need to push the artifact, but this is otherwise code-complete. Repo is [here](https://github.com/ZacSweers/auto-value-transient).

Resolves #198.